requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.10.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
azure-applicationinsights==0.1.1
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-search-documents==11.3.0
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cryptography==41.0.4
exceptiongroup==1.2.0
fastapi==0.103.2
Flask==2.3.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.0
frozenlist==1.4.1
greenlet==3.0.0
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
importlib_metadata==8.4.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.2
pluggy==1.3.0
portalocker==2.8.2
priority==2.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.1.0
pytest==7.2.2
python-dotenv
pytest-cov==4.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.4
typing==3.7.4.3
typing_extensions==4.8.0
urllib3==2.2.1
uvicorn==0.23.2
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.0
